implement the url for the json file
the json file can be specified as local path if it is in the same server as
the dashboard. Moreover it can be a common URL and sit in the remote
server. In order to read the json file in a remote server using AJAX, a CORS
(Cross-Origin Resouce Sharing) proxy is used to lift the restriction of cross
origin http requestes by browers.