implement the function that let users to load their own json files
A input menu with the name of "Browse" was added, so users can add
their own json files in their client side computer (i.e. the computer
in which the web browser is used to open the Unifed Dashboard in
the remote server such as lmt.ornl.gov).

Currently the json files in the client side computer is required to
be compatible with the CMEC JSON schema.