fix the error of no response when changing fixed dimension menus
When the dashboard is loaded or refreshed, there are no event
listeners on the fixed dimension menus. The event listseners
are added only when there is a change in the X or Y dimensions.

The error was corrected by adding the event listeners when the
dashboard is first loaded or refreshed.

Fixes: #1