Change X and Y dimensions of the default table when a JSON file is loaded
WHen a JSON file is loaded into memory, the dashboard can either
show a default table or an empty table (let users choose x, y and fixed
dimensions). Currently, the dashboard will show a default table with
model and metric as x and y dimensions respectively. But not all results
include model and metric dimensions. To fix this, the x and y dimensions
for default table will be the first and second dimnensions of json_structure
respectively. The other dimensions is the fixed dimensions and their default
values are the first values of their dimension variables.