fix two bugs and slightly adjust the layout
- Two bugs were fixed:
  - the changes of fixed dimensions do not update the object used
for the table after it goes through the normalization function.
  - changes in fixed dimensions will trigger the codes to create the
Tabulator instances and previous instances do not release memory
immediately. After several changes, the codes will use up all memory
and cause the dashboard crash.

The font variant of the bottom title of a table is changed to make the
text size same as that of the top title.

- Remove the "normalized" text and "Row" and "Column" are clear enough
as they are in the "scaling" section.