remove the indices and short_name from the statistic dimension
The statistic dimension is a special dimension in the previous
implementation of CMEC schema.  This PR will make it same as other
dimensions by removing the indices and short_name children dictionary
from it.

It is back compatible. Therefore, if users still use the dimension
named statistic and there are indice and short_name in it, UD can
handle it correctly.